+++ builder/datatypes/headerinfo.py
@@ -15,8 +15,10 @@ from builder.utils import assertion
 class HeaderInfo(object):
     ''' Header Info Data class.
     '''
-    def __init__(self, desc_chars: int, lines: int, papers: (int, float),
+    def __init__(self, total_chars: int, desc_chars: int,
+            lines: int, papers: (int, float),
             chapters: int, episodes: int, scenes: int, scodes: int):
+        self._total_chars = assertion.is_int(total_chars)
         self._desc_chars = assertion.is_int(desc_chars)
         self._lines = assertion.is_int(lines)
         self._papers = assertion.is_int_or_float(papers)
@@ -28,6 +30,10 @@ class HeaderInfo(object):
     #
     # property
     #
+
+    @property
+    def total_chars(self) -> int:
+        return self._total_chars
 
     @property
     def desc_chars(self) -> int: