+++ tests/datatypes/test_headerinfo.py
@@ -19,13 +19,14 @@ class HeaderInfoTest(unittest.TestCase):
         data = [
                 # (chars, lines, papers, chaps, epis, scenes, scodes,
                 #   exp_chars, exp_lines, exp_papers, exp_chaps, exp_epis, exp_scenes, exp_scodes)
-                (True, 1, 10, 100, 5, 5, 5, 5,
-                    1, 10, 100, 5,5,5,5),
+                (True, 9, 1, 10, 100, 5, 5, 5, 5,
+                    9, 1, 10, 100, 5,5,5,5),
                 ]
-        def checker(chars, lines, papers, chapters, episodes, scenes, scodes,
-                exp_chars, exp_lines, exp_papers, exp_chaps, exp_epis, exp_scenes, exp_scodes):
-            tmp = hd.HeaderInfo(chars, lines, papers, chapters, episodes, scenes, scodes)
+        def checker(total, chars, lines, papers, chapters, episodes, scenes, scodes,
+                exp_total, exp_chars, exp_lines, exp_papers, exp_chaps, exp_epis, exp_scenes, exp_scodes):
+            tmp = hd.HeaderInfo(total, chars, lines, papers, chapters, episodes, scenes, scodes)
             self.assertIsInstance(tmp, hd.HeaderInfo)
+            self.assertEqual(tmp.total_chars, exp_total)
             self.assertEqual(tmp.desc_chars, exp_chars)
             self.assertEqual(tmp.lines, exp_lines)
             self.assertEqual(tmp.papers, exp_papers)